+++ indra/llcorehttp/examples/http_texture_load.cpp
@@ -519,7 +519,7 @@ void WorkingSet::onCompleted(LLCore::HttpHandle handle, LLCore::HttpResponse * r
         {
             // More success
             LLCore::BufferArray * data(response->getBody());
-            mByteCount += data ? data->size() : 0;
+            mByteCount += data ? static_cast<long>(data->size()) : 0L;
             ++mSuccesses;
         }
         else
@@ -602,7 +602,7 @@ void WorkingSet::loadAssetUuids(FILE * in)
             mAssets.push_back(asset);
         }
     }
-    mRemaining = mLimit = mAssets.size();
+    mRemaining = mLimit = static_cast<int>(mAssets.size());
 }
